Pull out some comon code
