+++ packages/bodiless-navigation/src/Breadcrumbs/BreadcrumbStore/index.ts
@@ -1,1 +1,3 @@
 export * from './BreadcrumbStore bl-edit';
+export * from './BreadcrumbItem';
+export * from './types';
